chore(coin): move back cross to live-common
